requirements.txt
Jinja2
redis
Werkzeug
urllib3
mock
mockredispy
